Fix: replace Key? to Key due detailed error on description
Below it's the message Error from Dart due Key? declaration
---
A [Key] is an identifier for [Widget]s, [Element]s and [SemanticsNode]s.

A new widget will only be used to update an existing element if its key is the same as the key of the current widget associated with the element.

www.youtube.com/watch?v=kn0EOS-ZiIc

Keys must be unique amongst the [Element]s with the same parent.

Subclasses of [Key] should either subclass [LocalKey] or [GlobalKey].

See also:

[Widget.key], which discusses how widgets use keys.
This requires the 'non-nullable' language feature to be enabled.
Try updating your pubspec.yaml to set the minimum SDK constraint to 2.12.0 or higher, and running 'pub get'.